requirements.txt
python-rtmidi>=1.5.0
mido>=1.3.0
pyyaml>=6.0
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
httpx>=0.25.0
pynput>=1.7.6
